Do not upgrade packs pinned in the Require: section.
